safari-534.53-branch/LayoutTests/platform/mac/accessibility/aria-table-with-colspan-cells-expected.txt

Fremont
Leaves	Arrives	Route
7:00 AM	8:00 AM	FB1
7:00 AM	8:00 AM
7:00 AM	8:00 AM
This tests if an ARIA grid has a table cell with colspans, then cellForRowAndColumn still returns the correct cell.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS grid.cellForColumnAndRow(2, 0).role is 'AXRole: AXCell'
PASS grid.cellForColumnAndRow(2, 0).description is 'AXDescription: ColspanCell'
PASS grid.cellForColumnAndRow(2, 4).role is 'AXRole: AXCell'
PASS grid.cellForColumnAndRow(2, 4).description is 'AXDescription: RowspanCell'
PASS successfullyParsed is true

TEST COMPLETE
